# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for testing",
        "title": "GOLang API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "paths": {
        "/login": {
            "post": {
                "description": "User Login API",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login",
                        "name": "parameter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginModel"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/register": {
            "post": {
                "description": "User Register API",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "Register",
                        "name": "parameter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterModel"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/transfercredit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Transfer credit API",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfer Credit"
                ],
                "summary": "TransferCredit",
                "parameters": [
                    {
                        "description": "TransferModel",
                        "name": "parameter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.TransferModel"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/transfercredithistory": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Transfer credit API",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfer Credit"
                ],
                "summary": "TransferCreditHistory",
                "responses": {}
            }
        },
        "/users/userinfo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "User info API",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "UserInfo",
                "responses": {}
            }
        }
    },
    "definitions": {
        "auth.LoginModel": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "auth.RegisterModel": {
            "type": "object",
            "required": [
                "accountno",
                "fullname",
                "password",
                "username"
            ],
            "properties": {
                "accountno": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "user.TransferModel": {
            "type": "object",
            "properties": {
                "credit": {
                    "type": "integer"
                },
                "toAccountNo": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}